Z80Instruction 修正 - 型を Token へ変更- Token.Op を int へ変更

+++ parser/token.go
@@ -6,7 +6,7 @@ import "fmt"
 type Token struct {
 	Type    int
 	Literal string
-	Op      byte
+	Op      int
 	Line    int
 	Column  int
 }
